se agregan algunas excepciones

+++ src/main/java/com/example/parqueadero/repositorys/CarroRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CarroRepository extends JpaRepository<Carro , Integer> {
+public interface CarroRepository extends JpaRepository<Carro , Long> {
 }
